Use i32 for current user.

+++ app/src/audit/filter.rs
@@ -72,7 +72,7 @@ pub async fn filter(
     super::super::rls::set_row_level_security_user(&transaction, &current_user).await?;
 
     let user = queries::users::get()
-        .bind(&transaction, &(current_user.user_id as i32))
+        .bind(&transaction, &current_user.user_id)
         .one()
         .await?;
     let initials = crate::layout::initials(&user.email, user.first_name, user.last_name);
