+++ app/src/secrets/delete_secret.rs
@@ -43,7 +43,7 @@ pub async fn delete(
     queries::audit::insert()
         .bind(
             &transaction,
-            &(current_user.user_id as i32),
+            &current_user.user_id,
             &organisation_id,
             &AuditAction::DeleteSecret,
             &AuditAccessType::Web,
