+++ app/src/team/delete_member.rs
@@ -37,7 +37,7 @@ pub async fn delete(
     queries::audit::insert()
         .bind(
             &transaction,
-            &(current_user.user_id as i32),
+            &current_user.user_id,
             &delete_member.organisation_id,
             &AuditAction::CreateInvite,
             &AuditAccessType::Web,
