+++ app/src/vaults/index.rs
@@ -24,16 +24,12 @@ pub async fn index(
         .await?;
 
     let vaults = queries::vaults::get_all()
-        .bind(
-            &transaction,
-            &(current_user.user_id as i32),
-            &organisation_id,
-        )
+        .bind(&transaction, &current_user.user_id, &organisation_id)
         .all()
         .await?;
 
     let user = queries::users::get()
-        .bind(&transaction, &(current_user.user_id as i32))
+        .bind(&transaction, &current_user.user_id)
         .one()
         .await?;
     let initials = crate::layout::initials(&user.email, user.first_name, user.last_name);
